requirements/development.txt
Django==5.1.5
djangorestframework==3.15.2
python-dotenv==1.0.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
ruff==0.9.6
selenium==4.29.0
geckodriver-autoinstaller==0.1.0
psycopg2-binary==2.9.10
tornado==6.4.2
djangorestframework-simplejwt==5.5.0
